Mudancas no gradle

+++ lab3/app/src/classes/taxi/Location.java
@@ -23,7 +23,7 @@ public enum Location {
     public int getX(){
         return this.x;
     }
-    public int getY(){
+    public float getY(){
         return this.y;
     }
     public String getName(){
